Addition of modifyTransitionTable and addToTransitionTable Methods
We have added the modifyTransitionTable method and the addToTransitionTable
method. In addition we found a bug in our initTransitionTable method which
was not properly constructing the transition table to start with. This bug
has been fixed. We have also added a printTransitionTable Method. What
still needs to be done is to change the transitionTables type to
ArrayList<StateID[]> and to change anything else that is affected by this.