A Big Scare and Many Changes. All Resolved.
Found a problem with how we were moving. We did not understand what the tic
method was really doing. Each tic call is an actual move that you make.
Do to this we have edited: testConjecture, analyzeMove, run, and the class
ConjecturePathReturn. The class is now called ListAndBool. TestConjecture
now returns a ListAndBool that holds either the conjecturePath or a list of
moves that have been made that find the conjectured state to be false. This
path will be used in addPathToTransTable to add all of the new episodes
to the transition table.