Debugging in the Works
Changes were made to parts of the code for error checking and for dealing
with some out of bounds errors and null pointer exceptions. We now think
the problem that is occuring is that we may not be adding all of the rows
to the transitionTable, where we need to be. Need to run through with the
debugger to make sure all of the rows are being added correctly.