Debugging an Off By One Error.
We have cleaned up quite a few bugs in our program but we seem to still be
stuck with an off by one error somewhere in our code. We did manage to get
the program to run without a run time error but it appeared to be infinite
looping. See line 428 to see the commented out thing that we think was
causing an infinite loop. See my notes on my P drive do see a lot of the
changes that we made or notes made on today's work.