Added Comments To MakeMove Function
These comments specify things that need to be changed. We have looked over
some of the code that we have written and need to go through the work flow
of the program. Some of the methods that are in SmartAgent may become
superfluous because of the transitionTable Class. All emediate errors
due to the removal of the array list TransitionTable have been resolved.