Found More Problems
Since our last commit we think we have found some more problems in our code
specifically in analyzeMove in smartAgent when we went to try and abstract
parts of our code. I have added my notesAI.txt file which contains all of
my notes for the project to date so that they are easily accessible to
everyone so that they can see what we have done and what decisions we have
made and why. Easier than trying to fit everything into a commit message.