Re-Writing analyzeMove() and Notes
In the file temp.java, there is a version of analyzeMove() that is being
worked on. The idea is to re-write analyzeMove and abstract more out of it
to make it more readable and easier to debug should there be a problem.
This also serves as a way to go through the logic of it again to make sure
that there are as few logic errors as possible. Look at the Notes to get an

idea for one of the larger logic problems that we have discovered recently.