Added to test_smartAgent
Testing is still in progress. Found some confusion in the overloaded
checkIfEpisodeOccurred method as the overloaded version seemed
to have a different idea of itelf in the header than what it was doing.
This method only ever searches episodic memory, never any other list,
This was not clear from the comments, should be clearer now. All tests to
date pass.