Addition of new AnalyzeMove to SmartAgent and a Test Case for AnalyzeMove
The new AnalyzeMove() has been added to the smartAgent class and we are now
debugging analyzeMove(). It currently is passing the test Case but the test
case is not complete yet. The program is running however but it looks like
it might be infinte looping because the print out of states is getting up
to about 32762 since there is a print statement somewhere in the
transitionTable class. This may be due to a few changes in the analyzeMove
method that were made after meeting with Nux. Before it was failing the
test case. Will have to look into it more later.