Debugging
We have made a few changes to the code based on bugs that we found while
stepping through the program using the debugger. Some of these were: not
adding the goal state to the transition Table during initTransTable, or an
off by one error when building the conjecturePath. The code still needs
more debugging.