Added Changes to the Run Method and Contains Unknown Transitions
We changed the run method so that if you don't find a matching path, you
make a random move, and if you continue to not find a move these moves are
not added to the episodic memory, only the currentPath. We have also
changed the method containsUnknownTransitions(StateID goalState) so that it
now takes the goalState and that if it finds a row that has a move that
goes to the goal state, it considers that row in the transition Table
'complete'.